[REFACTOR] 폴더링
