+++ Dots/Dots/View/MyStrengthDetailView/Components/CustomDetailList.swift
