+++ Dots/Dots/View/MyStrengthDetailView/Components/StrengthNoteModal.swift
