+++ Dots/Dots/View/MyStrengthDetailView/Components/StrengthNoteViewModal.swift
