+++ Dots/Dots/View/SearchView/Modal/StrengthList.swift
